Added a new mixin and refactored tests to individual files
