+++ instructor/function_calls.py
@@ -82,140 +82,6 @@ class OpenAISchema(BaseModel):
             "input_schema": cls.model_json_schema(),
         }
 
-    def has_async_validators(self):
-        has_validators = (
-            len(self.__class__.get_async_validators()) > 0
-            or len(self.get_async_model_validators()) > 0
-        )
-
-        for _, attribute_value in self.__dict__.items():
-            if isinstance(attribute_value, OpenAISchema):
-                has_validators = (
-                    has_validators or attribute_value.has_async_validators()
-                )
-
-                # List of items too
-            if isinstance(attribute_value, (list, set, tuple)):
-                for item in attribute_value:
-                    if isinstance(item, OpenAISchema):
-                        has_validators = has_validators or item.has_async_validators()
-
-        return has_validators
-
-    @classmethod
-    def get_async_validators(cls):
-        validators = [
-            getattr(cls, name)
-            for name in dir(cls)
-            if hasattr(getattr(cls, name), ASYNC_VALIDATOR_KEY)
-        ]
-        return validators
-
-    @classmethod
-    def get_async_model_validators(cls):
-        validators = [
-            getattr(cls, name)
-            for name in dir(cls)
-            if hasattr(getattr(cls, name), ASYNC_MODEL_VALIDATOR_KEY)
-        ]
-        return validators
-
-    async def execute_field_validator(
-        self,
-        func: Any,
-        value: Any,
-        context: Optional[AsyncValidationContext] = None,
-        prefix=[],
-    ):
-        try:
-            if not context:
-                await func(self, value)
-            else:
-                await func(self, value, context)
-
-        except Exception as e:
-            prefix_path = f" at {'.'.join(prefix)}" if prefix else ""
-            return ValueError(f"Exception of {e} encountered{prefix_path}")
-
-    async def execute_model_validator(
-        self, func: Any, context: Optional[AsyncValidationContext] = None, prefix=[]
-    ):
-        try:
-            if not context:
-                await func(self)
-            else:
-                await func(self, context)
-
-        except Exception as e:
-            prefix_path = f" at {'.'.join(prefix)}" if prefix else ""
-            return ValueError(f"Exception of {e} encountered{prefix_path}")
-
-    async def get_model_coroutines(
-        self, validation_context: dict[str, Any] = {}, prefix=[]
-    ):
-        values = dict(self)
-        validators = self.__class__.get_async_validators()
-        model_validators = self.get_async_model_validators()
-        coros: list[Awaitable[Any]] = []
-        for validator in validators + model_validators:
-            # Model Validator
-            if not hasattr(validator, ASYNC_VALIDATOR_KEY):
-                validation_func, requires_validation_context = getattr(
-                    validator, ASYNC_MODEL_VALIDATOR_KEY
-                )
-                coros.append(
-                    self.execute_model_validator(
-                        validation_func,
-                        AsyncValidationContext(context=validation_context)
-                        if requires_validation_context
-                        else None,
-                        prefix,
-                    )
-                )
-            else:
-                fields, validation_func, requires_validation_context = getattr(
-                    validator, ASYNC_VALIDATOR_KEY
-                )
-                for field in fields:
-                    if field not in values:
-                        raise ValueError(f"Invalid Field of {field} provided")
-
-                    coros.append(
-                        self.execute_field_validator(
-                            validation_func,
-                            values[field],
-                            AsyncValidationContext(context=validation_context)
-                            if requires_validation_context
-                            else None,
-                            prefix=prefix + [field],
-                        )
-                    )
-
-        for attribute_name, attribute_value in self.__dict__.items():
-            # Supporting Sub Array
-            if isinstance(attribute_value, OpenAISchema):
-                coros.extend(
-                    await attribute_value.get_model_coroutines(
-                        validation_context, prefix=prefix + [attribute_name]
-                    )
-                )
-
-            # List of items too
-            if isinstance(attribute_value, (list, set, tuple)):
-                for item in attribute_value:
-                    if isinstance(item, OpenAISchema):
-                        coros.extend(
-                            await item.get_model_coroutines(
-                                validation_context, prefix=prefix + [attribute_name]
-                            )
-                        )
-
-        return coros
-
-    async def model_async_validate(self, validation_context: dict[str, Any] = {}):
-        coros = await self.get_model_coroutines(validation_context)
-        return [item for item in await asyncio.gather(*coros) if item]
-
     @classmethod
     def from_response(
         cls,
